% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_PlotPallete.R
\name{PlotPallete}
\alias{PlotPallete}
\title{PlotPallete}
\usage{
PlotPallete(colors.pallete = c(), dim.x = 4, dim.y = NULL,
  text.color = "white", ...)
}
\arguments{
\item{colors.pallete}{-- vector of strings with colors names}

\item{dim.x}{-- number of tiles on x axes}

\item{dim.y}{-- number of tiles on y axes}

\item{text.color}{-- color of color name label on tile}
}
\description{
PlotPallete
}
