% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleDensities.R
\name{rescaleDensitiesValues}
\alias{rescaleDensitiesValues}
\title{rescaleDensitiesValues}
\usage{
rescaleDensitiesValues(densities.list, density.rescale.fun = "numeric",
  density.rescale.fun.args = list(), ...)
}
\arguments{
\item{density.rescale.fun}{parameter, that defines a function used for rescaling signals in plots.
There are three built-in functions, that can be chosen:
(1) \code{'numeric'},
(2) \code{logarithmic} - with base defined in \code{density.rescale.fun.args} - default: \code{e = exp(1)}.
Function must be defined as a lambda construct \code{function(x, ...){...}}.}

\item{density.rescale.fun.args}{list of the arguments to defaults \code{density.rescale.fun}}
}
\description{
rescaleDensitiesValues
}
