% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleDensities.R
\name{rescaleSignalsValues.logarithmic}
\alias{rescaleSignalsValues.logarithmic}
\title{rescaleSignalsValues.logarithmic}
\usage{
rescaleSignalsValues.logarithmic(base = exp(1), ...)
}
\description{
rescaleSignalsValues.logarithmic
}
