% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_RescaleDensities.R
\name{rescaleSignalsValues.numeric}
\alias{rescaleSignalsValues.numeric}
\title{rescaleSignalsValues.numeric}
\usage{
rescaleSignalsValues.numeric(numeric.scale = 1, ...)
}
\description{
rescaleSignalsValues.numeric
}
