% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plots_ScatterBoxGGplot.R
\name{ScatterBoxplotGGplot}
\alias{ScatterBoxplotGGplot}
\title{ScatterBoxplotGGplot}
\usage{
ScatterBoxplotGGplot(data, x_, y_, point.size = 0.5, point.alpha = 0.5,
  point.scale = 0.25, scaled = TRUE)
}
\description{
ScatterBoxplotGGplot
}
