% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_plots_ScatterViolinGGplot.R
\name{ScatterViolinGGplot}
\alias{ScatterViolinGGplot}
\title{ScatterViolinGGplot}
\usage{
ScatterViolinGGplot(data, x_, y_, point.size = 0.5, point.alpha = 0.5,
  scaled = TRUE)
}
\description{
ScatterViolinGGplot
}
