% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_themes_ThemeSysbiosig.R
\name{theme_sysbiosig}
\alias{theme_sysbiosig}
\title{theme_sysbiosig}
\usage{
theme_sysbiosig(theme.base_size = 9, theme.base_family = "sans",
  theme.title_size = 12, theme.text_size = 9,
  theme.text.axes_size = 9, theme.text.strip_size = 9,
  theme.text.legend_size = 12, theme.margins = c(1, 1, 1, 1),
  legend.position = "right", ...)
}
\arguments{
\item{theme.base_size}{= 12,}

\item{theme.base_family}{= "sans",}

\item{theme.title_size}{= 36,}

\item{theme.text_size}{= 3*theme.title_size/4,}

\item{theme.margins}{= c(1,1, 1, 1),}

\item{legend.position}{= "right",}
}
\value{
ggthemes::theme_foundation object
}
\description{
This function specify theme used in the ITRC package
}
