% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics_ggplot_TilesGGPlot.R
\name{TilesGGplot}
\alias{TilesGGplot}
\title{TilesGGplot}
\usage{
TilesGGplot(data, x_, y_, response_, point.size = 0.5,
  point.alpha = 0.1, point.scale = 0.25, colors.limits = NULL)
}
\description{
TilesGGplot
}
